requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
plotly>=5.18.0
ta>=0.11.0
python-binance>=1.0.19
huggingface-hub>=0.20.0
pyarrow>=14.0.0
joblib>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0